added machine routes & refactored some shit
